Promote timed_step to being a general utility function
I was going to use it in the linter, but then decided not to; however,
moving it makes sense as it's of general utility.
